Renamed lucky_number package to luckynumber
