+++ src/main/kotlin/io/github/wulkanowy/api/service/HomepageService.kt
@@ -1,6 +1,6 @@
 package io.github.wulkanowy.api.service
 
-import io.github.wulkanowy.api.lucky_number.LuckyNumberResponse
+import io.github.wulkanowy.api.luckynumber.LuckyNumberResponse
 import io.reactivex.Single
 import retrofit2.http.GET
