+++ src/test/resources/io/github/wulkanowy/api/luckynumber/Index-withoutLuckyNumber.html
